requirements.txt
protobuf==3.1.0
tldextract==2.2.0
prometheus-client==0.4.0

gunicorn
flask
mock
coverage

numpy==1.15.2
torchtext==0.2.3
